backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
google-cloud-pubsub==2.18.4
pydantic==2.5.3
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
